Cleaned up usage, refactored code
