+++ main.go
@@ -12,18 +12,21 @@ import (
 	"github.com/hashicorp/terraform-provider-tfe/tfe"
 )
 
+const (
+	tfeProviderName = "registry.terraform.io/hashicorp/tfe"
+)
+
 func main() {
 	ctx := context.Background()
-	mainProvider := tfe.Provider().GRPCProvider
-	altProvider := tfe.AltServer
-
-	muxed, err := tfmux.NewSchemaServerFactory(ctx, mainProvider, altProvider)
+	muxed, err := tfmux.NewSchemaServerFactory(
+		ctx, tfe.Provider().GRPCProvider, tfe.ProviderServer,
+	)
 	if err != nil {
 		log.Println(err.Error())
 		os.Exit(1)
 	}
 
-	err = tf5server.Serve("registry.terraform.io/hashicorp/tfe", func() tfprotov5.ProviderServer {
+	err = tf5server.Serve(tfeProviderName, func() tfprotov5.ProviderServer {
 		return muxed.Server()
 	})
 	if err != nil {
