--- tfe/alt_data_source_state_output.go
@@ -1,47 +1,0 @@
-package tfe
-
-import (
-	"context"
-	"fmt"
-	"time"
-
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5/tftypes"
-)
-
-type dataSourceTime struct{}
-
-func (d dataSourceTime) ReadDataSource(ctx context.Context, req *tfprotov5.ReadDataSourceRequest) (*tfprotov5.ReadDataSourceResponse, error) {
-	state, err := tfprotov5.NewDynamicValue(tftypes.Object{
-		AttributeTypes: map[string]tftypes.Type{
-			"current": tftypes.String,
-			"id":      tftypes.String,
-		},
-	}, tftypes.NewValue(tftypes.Object{
-		AttributeTypes: map[string]tftypes.Type{
-			"current": tftypes.String,
-			"id":      tftypes.String,
-		},
-	}, map[string]tftypes.Value{
-		"current": tftypes.NewValue(tftypes.String, time.Now().Format(time.RFC3339)),
-		"id":      tftypes.NewValue(tftypes.String, "static_id"),
-	}))
-	if err != nil {
-		return &tfprotov5.ReadDataSourceResponse{
-			Diagnostics: []*tfprotov5.Diagnostic{
-				{
-					Severity: tfprotov5.DiagnosticSeverityError,
-					Summary:  "Error encoding state",
-					Detail:   fmt.Sprintf("Error encoding state: %s", err.Error()),
-				},
-			},
-		}, nil
-	}
-	return &tfprotov5.ReadDataSourceResponse{
-		State: &state,
-	}, nil
-}
-
-func (d dataSourceTime) ValidateDataSourceConfig(ctx context.Context, req *tfprotov5.ValidateDataSourceConfigRequest) (*tfprotov5.ValidateDataSourceConfigResponse, error) {
-	return &tfprotov5.ValidateDataSourceConfigResponse{}, nil
-}
